refactor: move extension function to outside of ViewModel
